Split ClipboardUtilities.cs, move the CF_HTML table parser to plain Core
Batch 5d of the Core split.

~330 of the file's 464 lines were a pure CF_HTML table parser with no
clipboard, WPF, WinRT or GDI+ dependency: BuildCfHtmlTable,
WrapHtmlFragmentAsCfHtml, ConvertHtmlToTabSeparated, ExtractHtmlFragment,
ParseHtmlTableToGrid, FindNextFreeColumnRange, ParseHtmlRowCells,
ParseSpanAttribute, CleanHtmlCellContent, and the MaxHtmlTableSpan constant.
Moved as-is to Text-Grab.Core/Utilities/CfHtmlTableUtilities.cs, keeping the
Text_Grab.Utilities namespace. WrapHtmlFragmentAsCfHtml and
ConvertHtmlToTabSeparated stayed internal - Text-Grab.Core already grants
InternalsVisibleTo to Text-Grab and Tests, so both the app's remaining
TryGetHtmlTableAsTabSeparated (now calling
CfHtmlTableUtilities.ConvertHtmlToTabSeparated) and the existing tests kept
working with no visibility change.

The clipboard-touching methods stayed in Text-Grab/Utilities/ClipboardUtilities.cs:
TryGetClipboardText, TryGetImageFromClipboard, GetBase64ClipboardContentAsImageSource,
ClipboardContainsBase64Image, CleanTeamsBase64Image, TryGetHtmlTableAsTabSeparated,
base64ImageExtension.

Separately, as instructed regardless of the split: line 64's
System.Windows.Forms.DataFormats.Bitmap (the file's only WinForms use) was
swapped for System.Windows.DataFormats.Bitmap - the identical string
constant, now sourced from WPF instead of WinForms.

Call sites updated: EditTextWindow.xaml.cs (BuildCfHtmlTable),
Tests/ClipboardUtilitiesTests.cs (BuildCfHtmlTable and
ConvertHtmlToTabSeparated, throughout), Tests/EditTextWindowSpreadsheetTests.cs
(the fully-qualified ConvertHtmlToTabSeparated call). None of these test
files moved to Tests.Core in this batch - that migration is Wave 7a's job,
not this one's.

Self-assembly-reflection audit: grepped both files for
GetExecutingAssembly/GetEntryAssembly/GetCallingAssembly, typeof(x).Assembly,
Assembly.GetName/.Location/.CodeBase, AssemblyMetadata, and
AppContext.BaseDirectory - zero matches.

Gate: dotnet build Tests/Tests.csproj -c Debug -p:Platform=x64 - clean
(0 errors; pre-existing MSB3277 assembly-conflict warnings only).
Tests.Core 514, Tests.Core.Windows 3, Tests 993 passed/7 skipped -
identical to baseline, total 1510 conserved.
